ES-529 Add CEC email template pre-filtering to reply functionality
