+++ app/controllers/support/management/email_templates_controller.rb
@@ -3,6 +3,7 @@ module Support
     class EmailTemplatesController < BaseController
       before_action :cec_template_group
       require "will_paginate/array"
+      include CecEmailTemplateFilters
 
       def index
         parser = Email::TemplateParser.new
@@ -64,24 +65,6 @@ module Support
         params.require(:email_template_form).permit(
           :group_id, :subgroup_id, :stage, :title, :description, :subject, :body, :blob_attachments, file_attachments: []
         ).merge(id: params[:id], agent: current_agent)
-      end
-
-      def filter_params
-        filters = params.fetch(:email_template_filters, {}).permit(:group_id, :remove_group, :remove_subgroup, :remove_stage, subgroup_ids: [], stages: [])
-
-        if (current_agent&.roles & %w[cec cec_admin]).any? && params[:email_template_filters].blank?
-          filters[:group_id] = @cec_group&.id
-          filters[:subgroup_ids] = [@dfe_subgroup&.id]
-        end
-
-        filters
-      end
-
-      def cec_template_group
-        @cec_group = Support::EmailTemplateGroup.find_by(title: "CEC")
-        if @cec_group.present?
-          @dfe_subgroup = Support::EmailTemplateGroup.find_by(title: "DfE Energy for Schools service", parent_id: @cec_group.id)
-        end
       end
 
       helper_method def portal_management_path
